+++ labo2/Squadron/ships/TIEInterceptor.cpp
@@ -11,13 +11,13 @@ using namespace std;
 
 unsigned TIEInterceptor::counter = 0;
 const string TIEInterceptor::MODEL_NAME = "TIE/IN";
-const unsigned TIEInterceptor::MAX_SPEED = 110;
+const double TIEInterceptor::MAX_SPEED = 110;
 const double TIEInterceptor::WEIGHT = 5;
 
 TIEInterceptor::TIEInterceptor() : Ship(++counter) {
 }
 
-unsigned TIEInterceptor::getMaxSpeed() const {
+double TIEInterceptor::getMaxSpeed() const {
 	return MAX_SPEED;
 }
