+++ labo2/Squadron/ships/cargo/Shuttle.cpp
@@ -11,14 +11,14 @@ using namespace std;
 
 unsigned Shuttle::counter = 0;
 const string Shuttle::MODEL_NAME = "Lamda-class shuttle";
-const unsigned Shuttle::MAX_SPEED = 54;
+const double Shuttle::MAX_SPEED = 54;
 const double Shuttle::WEIGHT = 360;
 const double Shuttle::MAX_CARGO = 80;
 
 Shuttle::Shuttle(double cargo) : CargoShip(++counter, cargo, MAX_CARGO) {
 }
 
-unsigned Shuttle::getMaxSpeed() const {
+double Shuttle::getMaxSpeed() const {
 	return MAX_SPEED;
 }
