+++ labo2/Squadron/ships/cargo/StarDreadnought.cpp
@@ -11,7 +11,7 @@ using namespace std;
 
 unsigned StarDreadnought::counter = 0;
 const string StarDreadnought::MODEL_NAME = "Super-class Star Destroyer";
-const unsigned StarDreadnought::MAX_SPEED = 40;
+const double StarDreadnought::MAX_SPEED = 40;
 const double StarDreadnought::WEIGHT = 9000000000;
 const double StarDreadnought::MAX_CARGO = 250000;
 
@@ -19,7 +19,7 @@ StarDreadnought::StarDreadnought(double cargo)
 	: CargoShip(++counter, cargo, MAX_CARGO) {
 }
 
-unsigned StarDreadnought::getMaxSpeed() const {
+double StarDreadnought::getMaxSpeed() const {
 	return MAX_SPEED;
 }
